fix database

+++ Areas/Api/RequestModels/BinhLuan/ThemBinhLuanModel.cs
@@ -5,7 +5,7 @@ namespace ChuyenDoiSoServer.Api.BinhLuan.RequestModel;
 public class ThemBinhLuanModel
 {
     public string NoiDung { get; set; }
-    public int TinTucId { get; set; }
-    public int UserId { get; set; }
-    public int? BinhLuanChaId { get; set; } = null;
+    public ulong TinTucId { get; set; }
+    public ulong UserId { get; set; }
+    public ulong? BinhLuanChaId { get; set; } = null;
 }
